Updated datatype

+++ src/main/java/com/example/TodoApiSpring/Todo.java
@@ -1,20 +1,20 @@
 package com.example.TodoApiSpring;
 
 public class Todo {
-    private int id;
+    private long id;
     private boolean completed;
     private String title;
     private int userId;
-    public Todo(int id, boolean completed, String title, int userId){
+    public Todo(long id, boolean completed, String title, int userId){
         this.id = id;
         this.completed = completed;
         this.title = title;
         this.userId = userId;
     }
-    public void setId(int id){
+    public void setId(Long id){
         this.id = id;
     }
-    public int getId(){
+    public long getId(){
         return id;
     }
